+++ spec/grape/endpoint_spec.rb
@@ -6,10 +6,13 @@ describe Grape::Endpoint do
 
   describe '#initialize' do
     it 'takes a settings stack, options, and a block' do
-      expect{ Grape::Endpoint.new(Grape::Util::HashStack.new, {
-        :path => '/',
-        :method => :get
-      }, &Proc.new{}) }.not_to raise_error
+      p = Proc.new {}
+      expect { 
+        Grape::Endpoint.new(Grape::Util::HashStack.new, {
+          :path => '/',
+          :method => :get
+        }, &p) 
+      }.not_to raise_error
     end
   end